+++ vllm/model_executor/layers/fused_moe/oracle/nvfp4.py
@@ -109,7 +109,7 @@ def backend_to_kernel_cls(
         return [FlashInferCuteDSLBatchedExperts]
 
     elif backend == NvFp4MoeBackend.FLASHINFER_B12X:
-        from vllm.model_executor.layers.fused_moe.experts.flashinfer_cutedsl_moe import (  # noqa: E501
+        from vllm.model_executor.layers.fused_moe.experts.flashinfer_b12x_moe import (  # noqa: E501
             FlashInferB12xExperts,
         )
